Small refactor of files to different folders
