+++ frontend/src/App.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter as Router } from "react-router-dom";
-import Navbar from "./navbar/Navbar";
+import Navbar from "./components/Navbar/Navbar";
 
 test("renders all links", () => {
   render(
