+++ frontend/src/components/Navbar/Navbar.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../styles/navbar.css";
+import "../../styles/navbar.css";
 
 function Navbar(): React.ReactElement {
   /* Add your routes and their names here and edit as needed. These were made with test data to make sure it was functioning */
