+++ frontend/src/pages/FileUploadPage/components/FileDropzoneControls.test.tsx
