+++ frontend/src/pages/FileUploadPage/components/FileDropzoneControls.tsx
